Modularized scroll fade and added it to folders
